+++ Planky/base/storage.py
@@ -6,6 +6,7 @@ from Planky.base.data.storageItem import StorageItem
 
 class ItemNotFound(Exception): pass
 class ItemExpired(Exception): pass
+class NoDefault: pass
 
 
 class Storage:
@@ -18,16 +19,19 @@ class Storage:
 
         }
 
-    def get(self, key: str):
+    def get(self, key: str, default: Any = NoDefault):
         '''
 
         :param key: Key of storage item
+        :param default: Default value
+        _if set will return default value instead of raising exception_
         :return: Value saved in storage
 
         :raises ItemNotFound: If item not found
         :raises ItemExpired: If item expired
         '''
         if key not in self._storage:
+            if default != NoDefault: return default
             raise ItemNotFound(f"Item {key} not found")
 
         item = self._storage.get(key)
@@ -35,6 +39,7 @@ class Storage:
         if item.expiration < 0 or item.expiration >= time():
             return item.value
         else:
+            if default != NoDefault: return default
             raise ItemExpired(f"Item {key} expired")
 
     def set(self, key: str, value: Any, lifetime: int = -1):